refactor: implement review
