+++ .gitignore
@@ -1,7 +1,10 @@
 # See http://help.github.com/ignore-files/ for more about ignoring files.
 
 .env
+
+# workspace caches
 .nx
+
 # compiled output
 dist
 tmp
