+++ packages/core/package.json
@@ -5,8 +5,7 @@
     "@code-pushup/models": "*",
     "@code-pushup/utils": "*",
     "@code-pushup/portal-client": "^0.1.2",
-    "chalk": "^5.3.0",
-    "zod": "^3.22.4"
+    "chalk": "^5.3.0"
   },
   "type": "commonjs",
   "main": "./index.cjs"
