+++ packages/core/src/index.ts
@@ -9,7 +9,7 @@ export {
 } from './lib/implementation/execute-plugin';
 export { collect, CollectOptions } from './lib/implementation/collect';
 export { upload, UploadOptions } from './lib/upload';
-export { GlobalOptions } from './lib/implementation/global-options';
+export { GlobalOptions } from './lib/types';
 export {
   collectAndPersistReports,
   CollectAndPersistReportsOptions,
