+++ packages/core/src/lib/types.ts
@@ -1,0 +1,6 @@
+export type GlobalOptions = {
+  //  'Show progress bar in stdout'
+  progress: boolean;
+  //  Outputs additional information for a run'
+  verbose: boolean;
+};
