+++ packages/core/src/lib/upload.ts
@@ -1,10 +1,10 @@
 import { uploadToPortal } from '@code-pushup/portal-client';
 import { CoreConfig, Report } from '@code-pushup/models';
 import { getLatestCommit, loadReport } from '@code-pushup/utils';
-import { GlobalOptions } from './implementation/global-options';
 import { jsonToGql } from './implementation/json-to-gql';
+import { GlobalOptions } from './types';
 
-export type UploadOptions = Pick<CoreConfig, 'upload' | 'persist'> &
+export type UploadOptions = Required<Pick<CoreConfig, 'upload' | 'persist'>> &
   GlobalOptions;
 
 /**
